# docs/openapi.generated.yaml
openapi: 3.1.0
info:
  title: Racing Point racecontrol API
  description: Generated from utoipa annotations. Admin-surface subset (Phase 445 D-06).
  contact:
    name: RacingPoint
    email: bono@racingpoint.in
  license:
    name: MIT
    identifier: MIT
  version: 1.0.0
paths:
...
  /api/v1/drivers:
    get:
      tags:
      - drivers
      operationId: list_drivers
      responses:
        '200':
          description: List of drivers (PII masked without manager role)
          content:
            application/json:
              schema: {}
        '401':
          description: Staff JWT required
      security:
      - staffJWT: []
...